Check operand number when operator clicked twice
Putting the check when a second number is clicked made it impossible to enter a number larger than single digit because it evaluated the result as soon as another number was entered. So I moved it to check for two numbers and and operation after a second operation is clicked.